Renaming Assignment3 to SnapFit (#17)
* Renaming Assignment3 to SnapFit

* format code and fix ktlint issues
