+++ app/src/androidTest/java/com/example/snapfit/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.assignment3
+package com.example.snapfit
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.assignment3", appContext.packageName)
+        assertEquals("com.example.snapfit", appContext.packageName)
     }
 }
