+++ app/src/main/java/com/example/snapfit/components/ExerciseCard.kt
@@ -22,9 +22,9 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.assignment3.R
 import com.example.compose.md_theme_light_primary
-import com.example.kotlinwithcompose.screens.LocalNavController
+import com.example.snapfit.R
+import com.example.snapfit.navigation.LocalNavController
 
 @Composable
 fun ExerciseCard() {
@@ -32,21 +32,21 @@ fun ExerciseCard() {
     var isToggled by remember { mutableStateOf(false) }
 
     Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(5.dp)
-            .border(5.dp, color = Color.Black, RoundedCornerShape(16.dp))
-            .padding(5.dp)
-            .background(
-                if (!isToggled) Color.White else md_theme_light_primary
-            )
-            .clickable { isToggled = !isToggled }
-
+        modifier =
+            Modifier
+                .fillMaxWidth()
+                .padding(5.dp)
+                .border(5.dp, color = Color.Black, RoundedCornerShape(16.dp))
+                .padding(5.dp)
+                .background(
+                    if (!isToggled) Color.White else md_theme_light_primary,
+                )
+                .clickable { isToggled = !isToggled },
     ) {
         Column(
             modifier = Modifier.weight(1f),
             verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
+            horizontalAlignment = Alignment.CenterHorizontally,
         ) {
             Text(
                 text = "Starfish",
