+++ app/src/main/java/com/example/snapfit/navigation/Routes.kt
@@ -1,4 +1,4 @@
-package com.example.assignment3.navigation
+package com.example.snapfit.navigation
 
 /**
  * The different possible routes of the navigation
@@ -6,7 +6,7 @@ package com.example.assignment3.navigation
  */
 sealed class Routes(val route: String) {
     object Main : Routes("MainScreenRoute")
-    
+
     object StartUp : Routes("StartUpRoute")
 
     object Login : Routes("LoginScreen")
