+++ app/src/main/java/com/example/snapfit/screen/ExercisesScreen.kt
@@ -1,4 +1,4 @@
-package com.example.assignment3.screen
+package com.example.snapfit.screen
 
 import ExerciseCard
 import androidx.compose.foundation.background
@@ -17,33 +17,33 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.compose.md_theme_light_primary
 
-
 @Composable
 fun ExercisesScreen() {
     Column(
-        modifier = Modifier
-            .padding(10.dp)
-            .verticalScroll(rememberScrollState())
+        modifier =
+            Modifier
+                .padding(10.dp)
+                .verticalScroll(rememberScrollState()),
     ) {
-       Column(Modifier.fillMaxWidth(),horizontalAlignment = Alignment.CenterHorizontally) {
-           Text(
-               text = "Yoga",
-               fontSize = 40.sp,
-               fontWeight = FontWeight.Bold,
-           )
-           Text(
-               text = "4 exercises",
-               fontSize = 20.sp,
-               fontWeight = FontWeight.Bold,
-           )
-           Text(
-               text = "Completed",
-               fontSize = 20.sp,
-               fontWeight = FontWeight.Bold,
-               textAlign = TextAlign.Center,
-               modifier = Modifier.background(color = md_theme_light_primary)
-           )
-       }
+        Column(Modifier.fillMaxWidth(), horizontalAlignment = Alignment.CenterHorizontally) {
+            Text(
+                text = "Yoga",
+                fontSize = 40.sp,
+                fontWeight = FontWeight.Bold,
+            )
+            Text(
+                text = "4 exercises",
+                fontSize = 20.sp,
+                fontWeight = FontWeight.Bold,
+            )
+            Text(
+                text = "Completed",
+                fontSize = 20.sp,
+                fontWeight = FontWeight.Bold,
+                textAlign = TextAlign.Center,
+                modifier = Modifier.background(color = md_theme_light_primary),
+            )
+        }
 
         ExerciseCard()
         ExerciseCard()
