+++ app/src/main/java/com/example/snapfit/screen/LoginScreen.kt
@@ -1,28 +1,15 @@
-package com.example.assignment3.screen
+package com.example.snapfit.screen
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
@@ -33,19 +20,13 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.semantics.SemanticsProperties.ImeAction
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.rememberNavController
-import com.example.assignment3.R
-import com.example.assignment3.navigation.Routes
-import com.example.kotlinwithcompose.screens.LocalNavController
+import com.example.snapfit.navigation.LocalNavController
+import com.example.snapfit.navigation.Routes
 
 /**
  * The about screen of the app, to display the use of the app.
@@ -55,14 +36,14 @@ import com.example.kotlinwithcompose.screens.LocalNavController
 fun LoginScreen() {
     val navController = LocalNavController.current
     Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(top = 80.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
+        modifier =
+            Modifier
+                .fillMaxSize()
+                .padding(top = 80.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
     ) {
         var username by remember { mutableStateOf("") }
         var password by remember { mutableStateOf("") }
-
 
         Text(
             text = "Loging",
@@ -72,49 +53,50 @@ fun LoginScreen() {
 
         Spacer(modifier = Modifier.height(70.dp))
 
-
         TextField(
             value = username,
             onValueChange = { username = it },
-            modifier = Modifier
-                .size(250.dp, 90.dp)
-                .padding(8.dp)
-                .border(3.dp, Color.Black)
-                .padding(8.dp),
+            modifier =
+                Modifier
+                    .size(250.dp, 90.dp)
+                    .padding(8.dp)
+                    .border(3.dp, Color.Black)
+                    .padding(8.dp),
             textStyle = TextStyle(fontSize = 16.sp),
-            keyboardOptions = KeyboardOptions.Default.copy(
-                keyboardType = KeyboardType.Text
-            ),
-            placeholder = { Text("Enter UserName", color = Color.Gray) }
+            keyboardOptions =
+                KeyboardOptions.Default.copy(
+                    keyboardType = KeyboardType.Text,
+                ),
+            placeholder = { Text("Enter UserName", color = Color.Gray) },
         )
-
 
         TextField(
             value = password,
             onValueChange = { password = it },
-            modifier = Modifier
-                .size(250.dp, 90.dp)
-                .padding(8.dp)
-                .border(3.dp, Color.Black)
-                .padding(8.dp),
+            modifier =
+                Modifier
+                    .size(250.dp, 90.dp)
+                    .padding(8.dp)
+                    .border(3.dp, Color.Black)
+                    .padding(8.dp),
             textStyle = TextStyle(fontSize = 16.sp),
-            keyboardOptions = KeyboardOptions.Default.copy(
-                keyboardType = KeyboardType.Password
-            ),
-            placeholder = { Text("Enter Password", color = Color.Gray) }
+            keyboardOptions =
+                KeyboardOptions.Default.copy(
+                    keyboardType = KeyboardType.Password,
+                ),
+            placeholder = { Text("Enter Password", color = Color.Gray) },
         )
-
 
         Button(
             onClick = {
                 navController.navigate(Routes.Main.route)
             },
-            modifier = Modifier
-                .padding(end = 8.dp)
-                .size(180.dp, 60.dp)
+            modifier =
+                Modifier
+                    .padding(end = 8.dp)
+                    .size(180.dp, 60.dp),
         ) {
             Text("Login")
         }
     }
 }
-
