+++ app/src/main/java/com/example/snapfit/screen/MainScreen.kt
@@ -1,6 +1,5 @@
-package com.example.assignment3.screen
+package com.example.snapfit.screen
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
