+++ app/src/main/java/com/example/snapfit/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.assignment3.ui.theme
+package com.example.snapfit.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
